Move rails specific to adaptor
